+++ cyberdrop_dl/scraper/crawlers/xenforo_crawler.py
@@ -284,7 +284,7 @@ class XenforoCrawler(Crawler):
     def is_attachment(self, link: URL) -> bool:
         if not link:
             return False
-        assert link.host
+        assert link.host, f"{link} has no host"
         parts = self.attachment_url_parts
         hosts = self.attachment_url_hosts
         return any(p in link.parts for p in parts) or any(h in link.host for h in hosts)
@@ -315,18 +315,15 @@ class XenforoCrawler(Crawler):
     async def handle_link(self, scrape_item: ScrapeItem, link: URL) -> None:
         if not link or link == self.primary_base_domain:
             return
-        try:
-            assert link.host
-            new_scrape_item = self.create_scrape_item(scrape_item, link)
-            if self.is_attachment(link):
-                return await self.handle_internal_link(new_scrape_item)
-            if self.primary_base_domain.host in link.host and self.stop_thread_recursion(new_scrape_item):  # type: ignore
-                origin = scrape_item.parents[0]
-                return log(f"Skipping nested thread URL {link} found on {origin}", 10)
-            new_scrape_item.set_type(None, self.manager)
-            self.handle_external_links(new_scrape_item)
-        except TypeError:
-            log(f"Scrape Failed: encountered while handling {link}", 40)
+        assert link.host, f"{link} has no host"
+        new_scrape_item = self.create_scrape_item(scrape_item, link)
+        if self.is_attachment(link):
+            return await self.handle_internal_link(new_scrape_item)
+        if self.primary_base_domain.host in link.host:  # type: ignore
+            origin = scrape_item.parents[0]
+            return log(f"Skipping nested thread URL {link} found on {origin}", 10)
+        new_scrape_item.set_type(None, self.manager)
+        self.handle_external_links(new_scrape_item)
 
     @error_handling_wrapper
     async def handle_internal_link(self, scrape_item: ScrapeItem) -> None: