+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented here. For more details, v
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [UNRELEASED]
+
+### Added
+
+- Iceyfile.com support
+- Support cookie extraction from Arc Browser, Lynx and W3M
+- `--exclude-files-with-no-extension` option: https://script-ware.gitbook.io/cyberdrop-dl/reference/configuration-options/settings/ignore_options#exclude_files_with_no_extension
+- `--print-stats` CLI only option
+
+### Removed
+
+- Cookie extraction support from Chrome on Windows
+
+### Fixed
+
+- Delete old database entries of previously unsupported URLs
+- Nudostar.tv crawler
+
 ## [6.4.0] - 2025-01-31
 
 ### Added